Rename openmc.data.container to openmc.data.function
