+++ openmc/data/__init__.py
@@ -3,7 +3,7 @@ from .neutron import *
 from .reaction import *
 from .ace import *
 from .angle_distribution import *
-from .container import *
+from .function import *
 from .energy_distribution import *
 from .product import *
 from .angle_energy import *
