+++ openmc/data/correlated.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import openmc.checkvalue as cv
 from openmc.stats import Tabular, Univariate, Discrete, Mixture, Uniform
-from .container import INTERPOLATION_SCHEME
+from .function import INTERPOLATION_SCHEME
 from .angle_energy import AngleEnergy
